requirements.txt
asgiref==3.9.1
certifi==2025.8.3
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.3
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fuzzywuzzy==0.18.0
gunicorn==23.0.0
idna==3.10
Levenshtein==0.27.1
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-Levenshtein==0.27.1
RapidFuzz==3.14.1
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
